+ Support for multi-passes materials added   Take in mind that this feature is still in development, so the result most likely will be very different from the actual wallpaper engine   The final approach might be different from what's currently in use, but It's just a nice proof of concept approach

+++ src/WallpaperEngine/Render/CCamera.h
@@ -14,7 +14,7 @@ namespace WallpaperEngine::Render
         CCamera (CScene* scene, Core::Scenes::CCamera* camera);
         ~CCamera ();
 
-        void setOrthogonalProjection (irr::u32 width, irr::u32 height);
+        void setOrthogonalProjection (irr::f32 width, irr::f32 height);
 
         irr::core::vector3df* getCenter ();
         irr::core::vector3df* getEye ();
